create shared about section
