+++ libs/application/templates/financial-statements-inao/src/fields/Shared/AboutOverview.tsx
@@ -1,0 +1,63 @@
+import React, { Fragment } from 'react'
+import { FieldBaseProps } from '@island.is/application/types'
+import { GridColumn, GridRow } from '@island.is/island-ui/core'
+import { formatPhoneNumber } from '@island.is/application/ui-components'
+import { FinancialStatementsInao } from '../../lib/utils/dataSchema'
+import { format as formatNationalId } from 'kennitala'
+import { m } from '../../lib/messages'
+import { ValueLine } from './ValueLine'
+
+export const AboutOverview = ({
+  answers,
+}: {
+  answers: FinancialStatementsInao
+}) => {
+  return (
+    <Fragment>
+      <GridRow>
+        <GridColumn span={['12/12', '6/12']}>
+          <ValueLine
+            label={m.nationalId}
+            value={
+              answers.about?.nationalId
+                ? formatNationalId(answers.about.nationalId)
+                : '-'
+            }
+          />
+        </GridColumn>
+        <GridColumn span={['12/12', '6/12']}>
+          <ValueLine label={m.fullName} value={answers.about.fullName} />
+        </GridColumn>
+      </GridRow>
+      <GridRow>
+        {answers.about.powerOfAttorneyName ? (
+          <GridColumn span={['12/12', '6/12']}>
+            <ValueLine
+              label={m.powerOfAttorneyName}
+              value={answers.about.powerOfAttorneyName}
+            />
+          </GridColumn>
+        ) : null}
+        {answers.about.powerOfAttorneyNationalId ? (
+          <GridColumn span={['12/12', '6/12']}>
+            <ValueLine
+              label={m.powerOfAttorneyNationalId}
+              value={formatNationalId(answers.about.powerOfAttorneyNationalId)}
+            />
+          </GridColumn>
+        ) : null}
+      </GridRow>
+      <GridRow>
+        <GridColumn span={['12/12', '6/12']}>
+          <ValueLine label={m.email} value={answers.about.email} />
+        </GridColumn>
+        <GridColumn span={['12/12', '6/12']}>
+          <ValueLine
+            label={m.phoneNumber}
+            value={formatPhoneNumber(answers.about.phoneNumber)}
+          />
+        </GridColumn>
+      </GridRow>
+    </Fragment>
+  )
+}
