+++ libs/application/templates/financial-statements-inao/src/fields/Shared/index.ts
@@ -1,2 +1,3 @@
 export { ValueLine } from './ValueLine'
 export { FileValueLine } from './FileValueLine'
+export { AboutOverview } from './AboutOverview'
