+++ libs/application/templates/financial-statements-inao/src/lib/utils/dataSchema.ts
@@ -369,7 +369,7 @@ export const dataSchema = z.object({
   equity,
   liability,
   attachments: z.object({
-    file: z.array(FileSchema).nonempty(),
+    file: z.array(FileSchema),
   }),
 })
